==> notebooks/01-consolidacao-metadados.ipynb <==
# Consolidação de Arquivos - Metadados



import os
import re
import time

import pandas as pd

def _obter_metadados_arquivos(caminho_arquivo: str) -> dict:
    metadados = os.stat(caminho_arquivo)
    return {
        "modo": metadados.st_mode,
        "inode": metadados.st_ino,
        "dispositivo": metadados.st_dev,
        "links": metadados.st_nlink,
        "uid": metadados.st_uid,
        "gid": metadados.st_gid,
        "tamanho": metadados.st_size,
        "ultimo_acesso": time.strftime("%d/%m/%Y %H:%M:%S", time.localtime(metadados.st_atime)),
        "ultima_modificacao": time.strftime("%d/%m/%Y %H:%M:%S", time.localtime(metadados.st_mtime)),
        "criacao": time.strftime("%d/%m/%Y %H:%M:%S", time.localtime(metadados.st_ctime))
    }

def metadados_arquivos(diretorio: str) -> dict:
    metadados = []
    for root, _dirs, files in os.walk(diretorio):
        for file in files:
            metadados_arq = _obter_metadados_arquivos(os.path.join(root, file))
            informacoes_arq = (
                {   "caminho": root,
                    "arquivo": file,
                    "extensao": re.sub("\.+", ".", file.split(".")[-1].lower())
                }
            )
            informacoes_arq.update(metadados_arq)
            metadados.append(informacoes_arq)
    return metadados

caminho_arquivos = "../arquivos"

df_metadados = pd.DataFrame(metadados_arquivos(caminho_arquivos))

df_metadados

(
    df_metadados.sort_values(by=["extensao","arquivo","caminho"])
    .to_csv("../resultados/metadados/consolidacao_metadados.csv", index=False, sep=";")
)
