==> notebooks/04-consolidacao-resultado.ipynb <==
# Consolidação de Arquivos - Etapa Final (Resultado)

import duckdb
import pandas as pd

df_consolidacao_estruturacao = pd.read_csv(
    "../resultados/consolidacao/consolidacao_estruturacao.csv",
    sep=";",
    dtype="str"
)

df_consolidacao_estruturacao.head(10)

#query = """
#SELECT * FROM df_consolidacao_estruturacao
#""" 

query = """
SELECT
        ID,
        MAX(CPF) CPF,
        MAX(Nome) Nome,
        MAX(Idade) Idade,
        MAX(Email) Email,
        MAX(Telefone) Telefone,
        MAX(Endereco) Endereco,
        MAX(Cargo) Cargo,
        MAX(Departamento) Departamento,
        MAX(Empresa) Empresa,
        MAX(Estado) Estado,
        MAX(Projeto) Projeto,
        MAX(Salario) Salario,
        MAX(Status) Status
FROM df_consolidacao_estruturacao
GROUP BY ID
ORDER BY ID
""" 

with duckdb.connect(database=":memory:", read_only=False) as conexao:
    df_consolidacao_final = conexao.sql(query).df()

df_consolidacao_final

(
    df_consolidacao_final.to_csv("../resultados/consolidacao_final.csv", index=False, sep=";")
)
